Dependency injection using picoContainer
