+++ pom.xml
@@ -41,6 +41,14 @@
     <version>4.1.1</version>
 </dependency>
 
+<!-- https://mvnrepository.com/artifact/io.cucumber/cucumber-picocontainer -->
+<dependency>
+    <groupId>io.cucumber</groupId>
+    <artifactId>cucumber-picocontainer</artifactId>
+    <version>7.8.0</version>
+</dependency>
+
+
   </dependencies>
 
   <build>
