+++ src/test/java/utils/TestContextSetup.java
@@ -1,0 +1,10 @@
+package utils;
+
+import org.openqa.selenium.WebDriver;
+
+public class TestContextSetup {
+
+	public WebDriver driver;
+	public String landingPageproductName;
+	
+}
